% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seal_fatness.R
\name{seal_fatness}
\alias{seal_fatness}
\title{Seal fatness}
\usage{
seal_fatness(seal_name, neck_rolls, girth)
}
\arguments{
\item{seal_name}{a character containing the name of the seal}

\item{neck_rolls}{a number containing the number of neck rolls the seal has}

\item{girth}{a character containing the estimated thiccness of the seal}
}
\value{

}
\description{
Prints a statment about how fat the seal is
}
