% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seal_squishyness.R
\name{seal_squishyness}
\alias{seal_squishyness}
\title{Title}
\usage{
seal_squishyness(fatness, friendly)
}
\arguments{
\item{fatness}{a character string of estimated seal fatness}

\item{friendly}{a number between 1 and 10 for how friendly this seal is}
}
\value{

}
\description{
Title
}
